pyproject.toml
[tool.poetry]
name = "konfik"
version = "1.0.0"
description = "The Strangely Familiar Config Parser"
authors = ["rednafi <redowan.nafi@gmail.com>"]

# New attributes
license = "MIT"
readme = "README.md"
homepage = "https://github.com/rednafi/konfik"
repository = "https://github.com/rednafi/konfik"
keywords = ["config", "configuration", "toml", "dotenv", "yaml"]
classifiers = [
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
]
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.6"
rich = "^8.0.0"
toml = "^0.10.1"
python-dotenv = "^0.14.0"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.1"
tox = "^3.20.0"
black = "^20.8b1"
pytest-cov = "^2.10.1"
isort = "^5.6.1"

[tool.poetry.scripts]
konfik = 'konfik.main:deploy_cli'

# linters
[tool.black]
l = 88

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
